Moved calculate_mb_toc_numbers to picard.disc.utils
