+++ picard/disc/eaclog.py
@@ -25,6 +25,8 @@
 
 import re
 
+from picard.disc.utils import calculate_mb_toc_numbers
+
 
 RE_TOC_TABLE_HEADER = re.compile(r""" \s*
     \s*.+\s+ \| # track
@@ -51,10 +53,6 @@ RE_TOC_TABLE_LINE = re.compile(r"""
 PREGAP_LENGTH = 150
 
 
-class NotSupportedTOCError(Exception):
-    pass
-
-
 def filter_toc_entries(lines):
     """
     Take iterator of lines, return iterator of toc entries
@@ -74,25 +72,6 @@ def filter_toc_entries(lines):
         if not m:
             break
         yield m.groupdict()
-
-
-def calculate_mb_toc_numbers(eac_entries):
-    """
-    Take iterator of toc entries, return a tuple of numbers for musicbrainz disc id
-    """
-    eac = tuple(eac_entries)
-    num_tracks = len(eac)
-    if not num_tracks:
-        raise NotSupportedTOCError("Empty track list: %s", eac)
-
-    expected_tracknums = tuple(range(1, num_tracks+1))
-    tracknums = tuple(int(e['num']) for e in eac)
-    if expected_tracknums != tracknums:
-        raise NotSupportedTOCError("Non-standard track number sequence: %s", tracknums)
-
-    leadout_offset = int(eac[-1]['end_sector']) + PREGAP_LENGTH + 1
-    offsets = tuple((int(x['start_sector']) + PREGAP_LENGTH) for x in eac)
-    return (1, num_tracks, leadout_offset) + offsets
 
 
 def toc_from_file(path):
